requirements.txt
faiss_cpu==1.11.0
numpy==2.2.6
protobuf==6.31.1
python-telegram-bot==22.1
sentence_transformers==4.1.0
